update api for combo building

+++ Back-end/src/main/java/com/example/system/model/combo/Material.java
@@ -22,7 +22,7 @@ public class Material {
     @Column(nullable = false, columnDefinition = "varchar(255)", unique = true)
     private String materialName;
     @Column(nullable = false)
-    private Double unitPrice;
+    private Long unitPrice;
     @ManyToOne
     @JoinColumn(name = "material_type_id")
     @JsonIgnore
